% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{br_singleblock_gamma_fit}
\alias{br_singleblock_gamma_fit}
\title{Gamma fit for br_singleblock data}
\format{
An object of class \link[=cumhist-class]{cumhist}
}
\usage{
br_singleblock_gamma_fit
}
\description{
Gamma fit for \link{br_singleblock} data using all defaults.
}
\details{
The call:
\code{br_singleblock_gamma_fit <-
        fit_cumhist(br_singleblock,
                    state = "State",
                    duration = "Duration")}
}
\seealso{
\code{\link[=fit_cumhist]{fit_cumhist()}}
}
\keyword{datasets}
