% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{br_singleblock_gamma_time_fit}
\alias{br_singleblock_gamma_time_fit}
\title{Gamma fit for br_singleblock data with log(Time) as fixed factor}
\format{
An object of class \link[=cumhist-class]{cumhist}
}
\usage{
br_singleblock_gamma_time_fit
}
\description{
Gamma fit for \link{br_singleblock} data with log(Time) as fixed
factor using all defaults.
}
\details{
The call:
\code{br_singleblock$LogTime <- log(br_singleblock$Time) \cr
      br_singleblock_gamma_time_fit <-
        fit_cumhist(br_singleblock,
                    state = "State",
                    duration = "Duration",
                    fixed_effects = "LogTime")}
}
\seealso{
\code{\link[=fit_cumhist]{fit_cumhist()}}
}
\keyword{datasets}
